Update 3 - Court
No court yet

+++ Data/Models/CarParts/EnginePart.cs
@@ -9,7 +9,7 @@
 
         public bool CheckFuel(Liquids carLiquids)
         {
-            byte fuelLevel;
+            int fuelLevel;
             carLiquids.CarLiquids.TryGetValue("Fuel", out fuelLevel);
 
             return fuelLevel >= 1;
